+++ obfuscator/src/test/java/by/radioegor146/VmTranslatorExecutionTest.java
@@ -40,10 +40,10 @@ public class VmTranslatorExecutionTest {
                     stack[sp++] = ins.operand;
                     break;
                 case VmOpcodes.OP_LOAD:
-                    stack[sp++] = locals[ins.operand];
+                    stack[sp++] = locals[(int) ins.operand];
                     break;
                 case VmOpcodes.OP_STORE:
-                    locals[ins.operand] = stack[--sp];
+                    locals[(int) ins.operand] = stack[--sp];
                     break;
                 case VmOpcodes.OP_ADD:
                     stack[sp - 2] += stack[sp - 1];
